+++ packages/dev/core/src/Misc/khronosTextureContainer2.ts
@@ -83,7 +83,7 @@ export class DefaultKTX2DecoderOptions {
         this._isDirty = true;
     }
 
-    private _useRGBAIfOnlyBC1BC3AvailableWhenUASTC = true;
+    private _useRGBAIfOnlyBC1BC3AvailableWhenUASTC?: boolean = true;
     /**
      * force a (uncompressed) RGBA transcoded format if transcoding a UASTC source format and only BC1 or BC3 are available as a compressed transcoded format.
      * This property is true by default to favor speed over memory, because currently transcoding from UASTC to BC1/3 is slow because the transcoder transcodes
@@ -93,7 +93,7 @@ export class DefaultKTX2DecoderOptions {
         return this._useRGBAIfOnlyBC1BC3AvailableWhenUASTC;
     }
 
-    public set useRGBAIfOnlyBC1BC3AvailableWhenUASTC(value: boolean) {
+    public set useRGBAIfOnlyBC1BC3AvailableWhenUASTC(value: boolean | undefined) {
         if (this._useRGBAIfOnlyBC1BC3AvailableWhenUASTC === value) {
             return;
         }
@@ -404,10 +404,7 @@ export class KhronosTextureContainer2 {
 
                         worker.addEventListener("error", onError);
                         worker.addEventListener("message", onMessage);
-
-                        if (KhronosTextureContainer2.DefaultDecoderOptions.isDirty) {
-                            worker.postMessage({ action: "setDefaultDecoderOptions", options: KhronosTextureContainer2.DefaultDecoderOptions._getKTX2DecoderOptions() });
-                        }
+                        worker.postMessage({ action: "setDefaultDecoderOptions", options: KhronosTextureContainer2.DefaultDecoderOptions._getKTX2DecoderOptions() });
 
                         const dataCopy = new Uint8Array(data.byteLength);
                         dataCopy.set(new Uint8Array(data.buffer, data.byteOffset, data.byteLength));
@@ -418,6 +415,9 @@ export class KhronosTextureContainer2 {
             });
         } else if (KhronosTextureContainer2._DecoderModulePromise) {
             return KhronosTextureContainer2._DecoderModulePromise.then((decoder) => {
+                if (KhronosTextureContainer2.DefaultDecoderOptions.isDirty) {
+                    KTX2DECODER.KTX2Decoder.DefaultDecoderOptions = KhronosTextureContainer2.DefaultDecoderOptions._getKTX2DecoderOptions();
+                }
                 return new Promise((resolve, reject) => {
                     decoder
                         .decode(data, caps)