+++ packages/dev/core/src/XR/native/nativeXRFrame.ts
@@ -80,6 +80,11 @@ export class NativeXRFrame implements XRFrame {
     public readonly getImageTrackingResults = (): XRImageTrackingResult[] => {
         return this._nativeImpl._imageTrackingResults ?? [];
     };
+
+    public getDepthInformation(view: XRView): XRCPUDepthInformation | undefined {
+        throw new Error("This function is not available in Babylon Native");
+        // return this._nativeImpl.getDepthInformation(view);
+    }
 }
 
 RegisterNativeTypeAsync("NativeXRFrame", NativeXRFrame);